Default .bikeys needed.
Removed this earlier, but had to make the change on mod updating side of things. Do no longer need to move bikeys to main keys directory when updating mods. This is done when the virtual server is made. Standard keys folder should be held only for official arma 3 bikeys (dlc)